Move fix mac names stuff to libraries
