+++ bin/gftools-fix-nameids.py
@@ -18,6 +18,9 @@ import argparse
 import os
 import tabulate
 from fontTools import ttLib
+from gftools.utils import has_mac_names
+from gftools.fix import drop_mac_names, drop_superfluous_mac_names
+
 
 parser = argparse.ArgumentParser(description='Print out nameID'
                                              ' strings of the fonts')
@@ -35,54 +38,6 @@ parser.add_argument('--drop-superfluous-mac-names', '-ms', default=False,
 parser.add_argument('--drop-mac-names', '-m', default=False,
                     action='store_true',
                     help='Drop all Mac name fields')
-
-
-def has_mac_names(ttfont):
-    """Check if a font has Mac names. Mac names have the following
-    field values:
-    platformID: 1, encodingID: 0, LanguageID: 0"""
-    for i in range(255):
-        if ttfont['name'].getName(i, 1, 0, 0):
-            return True
-    return False
-
-
-def drop_superfluous_mac_names(ttfont):
-    """Drop superfluous Mac nameIDs.
-
-    The following nameIDS are kept:
-    1: Font Family name,
-    2: Font Family Subfamily name,
-    3: Unique font identifier,
-    4: Full font name,
-    5: Version string,
-    6: Postscript name,
-    16: Typographic family name,
-    17: Typographic Subfamily name
-    18: Compatible full (Macintosh only),
-    20: PostScript CID,
-    21: WWS Family Name,
-    22: WWS Subfamily Name,
-    25: Variations PostScript Name Prefix.
-
-    We keep these IDs in order for certain application to still function
-    such as Word 2011. IDs 1-6 are very common, > 16 are edge cases.
-
-    https://www.microsoft.com/typography/otspec/name.htm"""
-    keep_ids = [1, 2, 3, 4, 5, 6, 16, 17, 18, 20, 21, 22, 25]
-    for n in range(255):
-        if n not in keep_ids:
-            name = ttfont['name'].getName(n, 1, 0, 0)
-            if name:
-                ttfont['name'].names.remove(name)
-
-
-def drop_mac_names(ttfont):
-    """Drop all mac names"""
-    for n in range(255):
-        name = ttfont['name'].getName(n, 1, 0, 0)
-        if name:
-            ttfont['name'].names.remove(name)
 
 
 def main():
